Rename File and create new alter buttons
